# Speed up SHA1 computation