The first commit is just a minor style fix that doesn't really warrant it's own PR.

The second commit speeds up SHA1 computation by ~50-60%. See the commit message for details. Unfortunately it does make the code a bit uglier than I initially expected because Rust doesn't support [destructuring assignments](https://github.com/rust-lang/rfcs/issues/372) (yet?). Still, even for an implementation like ring's, that favors simplicity over speed, the trade-off  is worth it in my opinion.

There's an alternative with much nicer code and the same speedup in [this branch](https://github.com/LingMan/ring/commits/sha1-speedup-crunchy). However, that turned out to increase the binary size by almost 390 KiB. Since ring is targeting small devices, that seamed unacceptable. Therefore I'm PRing this solution instead.

Before:
test digest::sha1::_1000     ... bench:       4,471 ns/iter (+/- 39) = 223 MB/s
test digest::sha1::_16       ... bench:         315 ns/iter (+/- 2) = 50 MB/s
test digest::sha1::_2000     ... bench:       8,903 ns/iter (+/- 208) = 224 MB/s
test digest::sha1::_256      ... bench:       1,452 ns/iter (+/- 14) = 176 MB/s
test digest::sha1::_8192     ... bench:      35,563 ns/iter (+/- 348) = 230 MB/s
test digest::sha1::block_len ... bench:         622 ns/iter (+/- 16) = 102 MB/s

After:
test digest::sha1::_1000     ... bench:       2,801 ns/iter (+/- 22) = 357 MB/s
test digest::sha1::_16       ... bench:         215 ns/iter (+/- 20) = 74 MB/s
test digest::sha1::_2000     ... bench:       5,556 ns/iter (+/- 86) = 359 MB/s
test digest::sha1::_256      ... bench:         922 ns/iter (+/- 35) = 277 MB/s
test digest::sha1::_8192     ... bench:      22,141 ns/iter (+/- 270) = 369 MB/s
test digest::sha1::block_len ... bench:         403 ns/iter (+/- 10) = 158 MB/s