@briansmith 2020-02-12
Thanks for doing this!

Are you interested in doing the same for the SHA-2 implementation?