@LingMan 2020-02-23
What exactly do you have in mind for SHA2? From a quick glance it doesn't appear to suffer from the same problem.